+++ src/App.tsx
@@ -6,8 +6,8 @@ import Header from './components/header/Header'
 
 import { Route, Routes } from 'react-router-dom'
 import { store } from './store'
-import Editor from './components/page/Editor/Editor'
-import Template from './components/page/Template/Template'
+import Editor from './page/Editor/Editor'
+import Template from './page/Template/Template'
 
 const App = () => {
     return (
